refactor(htmx): extract Requisition partials into app/routers/htmx/ package
First, lowest-blast-radius slice of the htmx_views.py split (Wave 2). Behavior
is identical — routes moved verbatim, same /v2 paths, same htmx-views tag.

- New package app/routers/htmx/:
  - _shared.py: cross-cutting helpers moved out of htmx_views (Vite manifest
    loader, _base_ctx(), _parse_date_safe()). Single source of truth; htmx_views
    re-imports them so its remaining routes are unchanged.
  - requisitions.py: the 12 Requisition partial routes (list, unified
    create/import modal + AI parse/save, detail shell, requirement add,
    search-all, detail tabs) + the AI customer lookup/quick-create the modal
    uses. Own APIRouter(tags=["htmx-views"]).
- main.py mounts htmx_requisitions_router alongside htmx_views_router (URLs
  unchanged).
- htmx_views.py re-imports requisitions_list_partial / requisition_tab because
  its offer/response routes re-render those partials; drops the now-unused
  parse_freeform_rfq / Path / etc. imports.
- Tests: retarget patch sites that followed moved symbols
  (parse_freeform_rfq -> htmx.requisitions, _vite_manifest -> htmx._shared).
- docs/APP_MAP_ARCHITECTURE.md: document the new package + extraction recipe.

Route parity verified: app.routes method+path dump is byte-identical before/after
(769 routes). Full suite green (21101 passed). pre-commit clean.
