+++ tests/test_req_import.py
@@ -47,7 +47,7 @@ def test_import_parse_html_returns_unified_modal(client, monkeypatch):
     async def mock_parse(text):
         return mock_result
 
-    monkeypatch.setattr("app.routers.htmx_views.parse_freeform_rfq", mock_parse)
+    monkeypatch.setattr("app.routers.htmx.requisitions.parse_freeform_rfq", mock_parse)
     resp = client.post(
         "/v2/partials/requisitions/import-parse",
         data={"name": "Test RFQ", "raw_text": "LM358DR 500 TI\nSTM32F407 100"},
